feat: Add `defineRaw` and `beforeCompile` options
Stylus `define()` method is now called with `raw = true` by default. This means objects literals passed via the `define` option will be converted to hashes instead of lists (expressions), which is much more useful.

If this functionality is undesired, the `defineRaw` option can be set to false. For more fine-grained control, the `beforeCompile` callback can be used to access the Stylus API directly.